multichannel analyser calibration
This analysis must be performed to ensure the multichannel analyser is calibrated. This calibration tells us whether there is any nonlinearity in the time scale of the time-to-amplitude converter.

Four delay values: (115.0 +/- 2.5) ns, (162.5 +/- 2.5) ns, (265 +/- 5) ns and (410 +/- 5) ns are used to shift the time scale forward in time, and then a least square fit is performed using curve_ft from scipy.optimize. The error in slope and intercept tells us if correction for nonlinearity is required.