Lifetime data analysis
Counts of decay event duration was measured experimentally, collected and displayed in a histogram. 

Python code was then used to analyse and extract the lifetime of positrons and positronium in varying pressures of CO2 gas. A major concern was the method used to reduce the impact of background signals.

We measured the vacuum spectra, and used the non-linear least squares optimisation function "curve_fit" from the SciPy module to find a scaling factor for the vacuum spectra, since non-vacuum and vacuum spectra data were taken for different durations.